Changing how hit detection works
Added a new Class called HitBox. I would like this class to replace what I was previously working on. HitBoxes should allow better checks for hit detection than the current method. Unfortunately they don't currently work the way I would like. Need to go through and see if this is a logic based problem or if it is that the visuals of the hitbox are not where the hitboxes truly are. I believe that it is the latter.