Transfering line on line based collision code into box on line based collision code
Changed the bottom and top line on line based collisions into box on line based collisions. I have kept the line on line based code as comments incase the boxes end up failing me. Figured out that the way the left/right hit detection works is the main problem with the current code for boxes. Will have to go over it and see if it can be fixed in some way. Preferably simplifying the slope equations with angles or radians or something.