Added animations and enums
Added more animations. Holding right/left currently causes it to go between left and right idle quickly. I found a way to get around this, but I think it will end up causing problems later so I'm going to leave it for now. I also added some enums to make the code easier to read. The id value of CollisionLine is now an enum. Additionally, the different animations are all enums as well.