Added a test map to start more testing
Added a test map to make testing more realistic. Some of the values inputted into the test map ended up being wrong and need to be fixed. Shouldn't have made it when I was so tired. A few collision errors were found, but I already know where the problems are for most of them.